requirements.txt
dash==2.18.2
dash_bootstrap_components==1.6.0
matplotlib==3.10.0
numpy==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
tf_nightly_intel==2.19.0.dev20241208
